Update Registration.java
Updated the imports.

Added a method to generate customer Ids.

Added usernameVerify method to check if username already exists in database when registering new account. Updated the actionPerformed method so that it used usernameVerify in IF statement to tell user if username already exists in database.

Added writeUserDataToCSV method to write user data to database.
 
Also added to actionPerformed is the use of generateRandomCustomerId method and then write everything all user information to CSV file.